lab-5.2-autogen-lab/requirements.txt
pyautogen==0.2.27
openai==1.23.6
python-dotenv==1.0.1
pytest==8.1.1
